+++ src/browserist/browser/wait/until/__main__.py
@@ -19,17 +19,17 @@ class WaitUntilDriverMethods(DriverMethods):
         self.text: WaitUntilTextDriverMethods = WaitUntilTextDriverMethods(browser_driver, settings)
         self.url: WaitUntilUrlDriverMethods = WaitUntilUrlDriverMethods(browser_driver, settings)
 
-    def element_disappears(self, xpath: str, timeout: int = timeout.DEFAULT) -> None:
+    def element_disappears(self, xpath: str, timeout: float = timeout.DEFAULT) -> None:
         """Wait until element doesn't exist."""
 
         wait_until_element_disappears(self._driver, xpath, timeout)
 
-    def images_have_loaded(self, xpath: str, timeout: int = timeout.DEFAULT) -> None:
+    def images_have_loaded(self, xpath: str, timeout: float = timeout.DEFAULT) -> None:
         """Wait until element doesn't exist. The image XPath can target one or more images."""
 
         wait_until_images_have_loaded(self._driver, xpath, timeout)
 
-    def number_of_window_handles_is(self, timeout: int = timeout.DEFAULT) -> None:
+    def number_of_window_handles_is(self, timeout: float = timeout.DEFAULT) -> None:
         """Wait until number of window handles is."""
 
         wait_until_number_of_window_handles_is(self._driver, timeout)
