+++ src/browserist/browser/wait/until/page_title/contains.py
@@ -6,7 +6,7 @@ from .....constant import timeout
 from .....exception.timeout import WaitForPageTitleToChangeTimeoutException
 
 
-def wait_until_page_title_contains(driver: object, page_title_fragment: str, timeout: int = timeout.DEFAULT) -> None:
+def wait_until_page_title_contains(driver: object, page_title_fragment: str, timeout: float = timeout.DEFAULT) -> None:
     try:
         WebDriverWait(driver, timeout).until(EC.title_contains(page_title_fragment))  # type: ignore
     except TimeoutException:
