+++ src/browserist/browser/wait/until/url/equals.py
@@ -7,7 +7,7 @@ from .....exception.timeout import WaitForUrlTimeoutException
 from .....model.type.url import URL
 
 
-def wait_until_url_equals(driver: object, url: str, timeout: int = timeout.DEFAULT) -> None:
+def wait_until_url_equals(driver: object, url: str, timeout: float = timeout.DEFAULT) -> None:
     url = URL(url)
     try:
         WebDriverWait(driver, timeout).until(EC.url_matches(url))  # type: ignore
